+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Updates
 
 - Add [CHANGELOG.md](./CHANGELOG.md)
+- Add [Tuple](./src/types/tuple.d.ts) type
 
 ## [1.0.2](https://github.com/havelessbemore/munkres/compare/v1.0.1...v1.0.2) (2024-03-31)